redict_be/requirements.txt
fastapi
uvicorn
requests
python-multipart
pytubefix
openai-whisper
pydantic
pydantic-settings